do some housekeeping
