+++ constants.js
@@ -1,0 +1,22 @@
+export const initialValues = {
+  enemyVelocity: 0.5,
+  powerUpVelocity: 1,
+  activeBug: 0,
+  randomspawncounter: 0,
+  topRow: {
+    0: null,
+    1: null,
+    2: null,
+    3: null,
+    4: null,
+    5: null,
+    6: null,
+    7: null,
+  },
+  sumValueOfEnemies: 0,
+  scale: 0.5,
+};
+
+export const TOP_BUFFER = 50;
+export const LEFT_BUFFER = 25;
+export const RIGHT_BUFFER = 15;
