+++ index.html
@@ -19,7 +19,7 @@
     <script src="https://cdn.jsdelivr.net/npm/phaser@3.16.2/dist/phaser.min.js"></script>
 
 
-  <script defer src="2048Invaders.js"></script>
+  <script defer src="2048Invaders.js" type="module"></script>
   </div>
 </body>
 </html>
